requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.2
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
crypto==1.4.1
cryptography==37.0.2
discord.py==1.7.3
dnspython==2.2.1
future==0.18.2
h11==0.13.0
idna==3.3
keyboard==0.13.5
multidict==6.0.2
Naked==0.1.31
outcome==1.1.0
pefile==2021.9.3
pycparser==2.21
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
pymongo==4.1.1
pynput==1.7.6
pyOpenSSL==22.0.0
PySocks==1.7.1
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
selenium==4.1.5
shellescape==3.8.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.1.1
urllib3==1.26.9
wsproto==1.1.0
yarl==1.7.2
